+++ src/main/java/se/rimevel/FeudalFunctions/modules/smithing/blocks/BlockForge.java
@@ -124,11 +124,11 @@ public class BlockForge extends BlockContainerBase implements ITemperatureModifi
 	}
 	
 	@Override
-	public int getTempMod()
+	public float getTempMod()
 	{
 		if(this.lightValue > 0)
 		{
-			return 2;
+			return 2.0F;
 		}
 		return 0;
 	}
