+++ src/main/java/se/rimevel/FeudalFunctions/modules/survival/interfaces/ITemperatureModifier.java
@@ -2,5 +2,5 @@ package se.rimevel.FeudalFunctions.modules.survival.interfaces;
 
 public interface ITemperatureModifier
 {
-	public int getTempMod();
+	public float getTempMod();
 }
